[module] Verplaats documenten naar een module
