+++ htdocs/.htaccess
@@ -24,7 +24,7 @@ RewriteRule ^wachtwoord			index.php?c=Login [L]
 RewriteRule ^account			index.php?c=Login [L]
 
 RewriteRule ^eetplan			index.php?m=eetplan&c=Eetplan	[QSA,L]
-RewriteRule ^documenten			index.php?c=Documenten	[QSA,L]
+RewriteRule ^documenten			index.php?m=documenten&c=Documenten	[QSA,L]
 RewriteRule ^bibliotheek		index.php?c=Bibliotheek [QSA,L]
 RewriteRule ^mededelingen       index.php?c=Mededelingen [QSA,L]
 RewriteRule ^peilingen          index.php?c=Peilingen [QSA,L]
